requirements.txt
asgiref==3.6.0
bidict==0.22.1
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.6
idna==3.4
mysql-connector-python==8.0.32
mysqlclient==2.1.1
protobuf==3.20.3
python-engineio==4.3.4
python-socketio==5.7.2
requests==2.28.2
requests-futures==1.0.0
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.14
yfinance
django-cors-headers